pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "all-money-back-me-home"
version = "0.1.0"
description = "Crypto futures auto-trading + multi-market indicator monitoring"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # Trading
    "freqtrade>=2024.1",
    "ccxt>=4.0.0",

    # HTTP / async
    "httpx>=0.27.0",

    # Data & Analysis
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "TA-Lib>=0.4.0",
    "yfinance>=0.2.0",

    # Taiwan stocks
    "twstock>=1.4.0",

    # Configuration
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",

    # MCP Server
    "fastmcp>=2.0.0",

    # Notifications
    "python-telegram-bot>=20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
